# pmap get_state returns interpolated state