--- a/src/orbitdiagrams/poincare.jl
+++ b/src/orbitdiagrams/poincare.jl
@@ -271,7 +271,6 @@ function Base.show(io::IO, pmap::PoincareMap)
     println(io, "Iterator of the Poincaré map")
     println(io,  rpad(" rule f: ", 14),     DynamicalSystemsBase.eomstring(pmap.integ.f.f))
     println(io,  rpad(" hyperplane: ", 14),     pmap.planecrossing.plane)
-    print(io,  rpad(" selection: ", 14),     pmap.i)
 end