--- a/test/orbitdiagrams/poincare_tests.jl
+++ b/test/orbitdiagrams/poincare_tests.jl
@@ -66,18 +66,18 @@ println("\nTesting poincare sections...")
     end
     @testset "Henon Heiles iterated Poincaré map" begin
         ds = Systems.henonheiles([0, 0.483000, 0.278980390, 0] )
-        pmap = poincaremap(ds, (2, 0.0), 20.0; direction=+1, idxs=[1,2])
+        pmap = poincaremap(ds, (2, 0.0); direction=+1)
         next = step!(pmap)
         @test abs(next[2]) < 1e-6
-        # Test that iteration terminates fast
-        for i in 1:50; step!(pmap); end
-        @test isnothing(step!(pmap))
         # Test that we can re-init
         reinit!(pmap, ds.u0)
         next2 = step!(pmap)
         @test !isnothing(next2)
         @test abs(next2[2]) < 1e-6
         @test next[1] ≈ next2[1] atol = 1e-3
+        # Test an ill defined hyperplane
+        pmap = poincaremap(ds, (2, 5.0), 100.0; direction=+1)
+        @test isnothing(step!(pmap))
     end
 end