This is an attempt to solve the problem of the Poincaré maps. In this PR the state returned by `get_state` is the state interpolated  on the Poincaré map. 

There is a small impact on the performance that I still don't know how to solve in `step!` (there is a `if else`). Tests are Ok.