@Datseris 2021-06-24
What's the performance regression when calling `poincaresos` as in the test suite?