@awage 2021-06-25
No impact on `poincaresos` but a small dent in performance for `step!` :

```
Master: 
julia> @btime psos = poincaresos(ds, (2, 0.0), 1000.0, direction = +1)
1.921 ms (2828 allocations: 253.23 KiB)

julia> @btime  for i in 1:100000; step!(pmap); end
  7.273 ms (500000 allocations: 15.26 MiB)

PR: 
  julia> @btime psos = poincaresos(ds, (2, 0.0), 1000.0, direction = +1)
  1.913 ms (2828 allocations: 253.23 KiB)

  julia> @btime  for i in 1:100000; step!(pmap); end
  7.614 ms (500000 allocations: 15.26 MiB)
```