@Datseris 2021-06-25
Hm this change in allocations is huge. You don't see its impact on performance because I assume you are using a low dimensional out of place system to do the calculations. For higher dimensional systems this would make a big difference.

In the first place, does this fix the problems we had with the basins of attraction?