@awage 2021-06-25
The change is big from `poincaresos` to `step!`. This PR doesn't add any allocation when compared to master version. 

This fix the problem with `basin_of_attraction`. What I though was a fantastic example of basins with the Rikitake ODE is in fact lame and boring: 

```
ds = Systems.rikitake(μ = 0.47, α = 1.0)
plane = (3, 0.0)
pmap = poincaremap(ds, (3, 0.), Tmax=1e6;
    idxs = 1:2, rootkw = (xrtol = 1e-12, atol = 1e-12), reltol=1e-9
)
xg = yg = range(-6.,6.,length=200)
basin, attractors = basins_of_attraction((xg, yg), pmap)

attractors[1]
3-dimensional Dataset{Float64} with 2 points
 -2.20231   1.05136   1.59021e-9
  2.36865  -1.33907  -4.63543e-10
```

The problem with this parameter is that there is long transient before we get to the attractor. It fools the algorithm that thinks he has found different attractors when you compute basisns without `pmap`. It can be avoided with `mx_chk_bas` or `mx_chk_att` but `pmap` requires less tweaking. This Rikitake is a tough guy. 

![imagen](https://user-images.githubusercontent.com/3494384/123407666-7ccf8c80-d5ac-11eb-8604-9cc2b6a441f8.png)