@awage 2021-06-25
I spotted a problem. `step!` returns `nothing` when the map is not well defined while `get_state` always returns a vector. We should be able to detect this in the algorithm. I have to rethink this PR. `pmap` is a headache but I think it is a valuable tool.