@awage 2021-06-30
While I am not very satisfied with the solution it does not affect the performance and fix the problem of tracking the trajectory on the plane. 

```
Master: 
@btime psos = poincaresos(ds, (2, 0.0), 1000.0, direction = +1)
  1.721 ms (2828 allocations: 253.23 KiB)

@btime  for i in 1:100000; step!(pmap); end
 6.384 ms (500000 allocations: 15.26 MiB)

PR: 
@btime psos = poincaresos(ds, (2, 0.0), 1000.0, direction = +1)
  1.718 ms (2828 allocations: 253.23 KiB)

@btime  for i in 1:100000; step!(pmap); end
  6.324 ms (500000 allocations: 15.26 MiB)
```

`Thomas_Cyclical` system is very sharp with `pmap`
![imagen](https://user-images.githubusercontent.com/3494384/123925673-6f8b1700-d98b-11eb-899d-feee70d6bc51.png)