@awage 2021-06-30
Also I found a bug in `check_next_state!` that should be pushed in master. If not in this PR in another.