@awage 2021-07-01
Hi!  Here is a quick summary of the reason of the changes: 
* `step!` returns `nothing` when the trajectory does not return to the map after `Tmax`. But if `get_state` returns `nothing` we have a problem with  `view` in `get_projected_state` that is expecting a vector.  
* The trick is to return a vector of `NaN` in `get_state` if the orbit does not return. This way we can check in `_identify_basin_of_cell!` if the orbit is off the plane (`get_state` is used to get the full state of the system). 
* This has no impact on the behavior of pmap since `step!` and `poincaresos` will use the low level function and return the interpolated projected state.

I think the fundamental problem is that `step!` does not behave in the same way for `PoincaŕeMap` and for `ContinuousDynamicalSystems`. In the second case you never get `nothing` and there is no type conflict. 

Another solution is to change the behavior of `step!` in order to be consistent with `get_state` so that they return both `NaN` or something characteristic that makes clear that the trajectory does not return to the map.