@Datseris 2021-07-03
BTW I think the current solution is just fine, I don't think it is more useful to make the return type a vector of NaN or something like that.