@Datseris 2021-07-03
Okay, I'm happy with the final version. One thing I found clunky and unintuitive is why the `PoincareMap` was having projection indices, given that it operates on the full system state. It made the remaining code confusing, and the interaction with the basins of attraction even more confusing. For example, the old code would find as attractors 3-dimensional points, even with the PoincareMap having `idxs = 1:2`. I thought that was weird. So I completely removed `idxs` from `PoincareMap`.

The other thing I did was change the logic of `Tmax` in `PoincareMap`. Previously, it would count time from the creation of the map. Now it instead counts time from the start of `step!`, which I think is more sensible and will better allow longer computation times without worrying whther the iteration exceeded `Tmax` even for a well-defined hyperplane.