@awage 2021-07-04
Awesome!! 

I found an example to compare the output of the algorithm with a published result (Sprott again). I am preparing a PR with this system: 
```
using DynamicalSystems
@inline @inbounds function dl_lorenz(u, p, t)
    R = p[1];
    x = u[1]; y = u[2]; z = u[3];
    dx = y - x
    dy = -x*z
    dz = x*y - R
    return SVector{3}(dx, dy, dz)
end
R=4.7;
ds = ContinuousDynamicalSystem(dl_lorenz, rand(3), [R])
xg = yg = range(-10.,10.,length=500)
pmap = poincaremap(ds, (3, 0.), Tmax=1e2)
basin,attractors = basins_of_attraction((xg,yg), pmap)
plot(xg,yg,basin',seriestype=:heatmap)
```

Original in J. C. Sprott,  Simplest Chaotic Flows with Involutional Symmetries, Int. Jour. Bifurcation and Chaos 24, 1450009 (2014):

![imagen](https://user-images.githubusercontent.com/3494384/124399637-fb4ecb80-dd1c-11eb-9681-4727dbd5258a.png)

Computed with `DynamicalSystems.jl` (30 seconds for 500x500!!) 
![imagen](https://user-images.githubusercontent.com/3494384/124399738-b5dece00-dd1d-11eb-912f-838e46fd91ab.png)